refactor(FE): PageTemplate 및 관련 컴포넌트 리팩토링 및 스크롤 상태 관리 제거
- PageTemplate 컴포넌트에서 pageName prop 제거 및 스크롤 상태 관련 코드 삭제
- App 컴포넌트에서 스크롤 방향 추적 로직 제거 및 SubHeader 컴포넌트 추가
- 여러 페이지에서 PageTemplate 사용 시 pageName prop을 제거하여 일관성 있는 구조로 변경
- Header, SubHeader 컴포넌트에서 Redux 스크롤 상태 대신 커스텀 훅을 사용하여 스크롤 방향 관리
- Redux 스크롤 상태 슬라이스 삭제
