+++ front/src/pages/Notification.tsx
@@ -14,7 +14,7 @@ export default function Notification() {
   });
 
   return (
-    <PageTemplate pageName="모든 알림">
+    <PageTemplate>
       <div className="max-w-xl mx-auto bg-background rounded-lg overflow-hidden">
         {notifications ? (
           notifications.map((notification, index) => (
