+++ front/src/pages/WriteReview.tsx
@@ -3,7 +3,7 @@ import ReviewForm from "@/widgets/ReviewForm";
 
 export default function WriteReview() {
   return (
-    <PageTemplate pageName="리뷰 작성">
+    <PageTemplate>
       <ReviewForm />
     </PageTemplate>
   );
