--- front/src/state/store/scrollStateSlice.ts
@@ -1,16 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-const scrollStateSlice = createSlice({
-  name: "scrollState",
-  initialState: {
-    isScrollingUp: true,
-  },
-  reducers: {
-    setScrollState: (state, action) => {
-      state.isScrollingUp = action.payload;
-    },
-  },
-});
-
-export const { setScrollState } = scrollStateSlice.actions;
-export default scrollStateSlice.reducer;
